requirements.txt
flask>=3.0.0
typing_extensions>=4.8.0
viam-sdk>=0.24.0

